move some tests to test_input_validation
